Revised luck timeouts. With the recent changes to how wishing impacts luck, the luck timeout table needed a bit of adjusting. With the current table, there's really only a spread of -1 to +10 luck covered. But, with the new system, the entire range from -10 luck to +10 luck needs timeouts. If a player gets their base luck down to -10, it's possible to drive it back up to +10 via unicorn gem luck or altar sacrificing.
I also thought the UnNetHack system of timeouts was a bit too generous. Under that system, it would take almost 100000 turns to timeout from +3 luck to 0 luck, and almost 200000 turns to go from +10 to 0.

I have revised it to be a bit more impactful, but still have fairly generous timeouts at the lower levels. To go from +1 luck to 0 is about 10000 turns, and I used this as the baseline.

- To go from +3 to 0 is about 35000 turns.
- To go from +6 to 0 is about 51000 turns.

If a player has a very low base luck (due to wish luck penalties), the timeouts can get quite short near the upper levels, but this is by design, otherwise the penalty is too easy to waive.