Shield properties can deal damage during power shield or shield bashing.
This is something that sounded really neat for the power shield tech, but after getting the groundwork done, I decided to throw it in for the regular shield bashing attacks too. It occurs on 25% of shield bashes, so not too spammy but still adding some significant damage.

The sleep and disease properties felt too weird for shields to inflict damage so I left those out.