Potion of amnesia tweaks.
When dipping negatively enchanted items, the enchantment will be cleared and reset to 0. Before it would drain positive enchantment and leave negative enchantment alone.

Also cleaned up some other loose ends. The behavior when dipping amnesia into water wasn't working correctly. There was also an artifact dip check that seemed unnecessary so I removed that.

An important tweak was if a player purposely dips an item with a property into amnesia - it will always remove all the properties on the item. This is to enable the player to easily remove harmful props. The chance of removing properties when hit by a thrown potion is still 1 in 13.

Amnesia will also now turn a crysknife into a worm tooth.