Attempting to eat phoenix eggs triggers them to hatch.
This is in response to mobileuser's observation that you can simply eat them to prevent them from endlessly regenerating. It feels a bit too easy for a player to thwart the key mechanic of this monster by just eating it - so let's up the ante a bit.

Currently you can cancel any egg (including phoenix eggs) by dipping them in amnesia potions. You can also bury a phoenix egg to dispose of it. In the next commit, I'm going to address canceling of eggs via wand or spell of cancellation, which currently doesn't seem to actually cancel eggs.