Phoenix egg tweaks. In the spirit of making phoenixes a recurring force to be reckoned with, they will hatch if the player attempts to eat, bury, or confine them in a container. In the previous commit, I allowed their eggs to be canceled, but I modified the chance to roll against the innate MR of the phoenix - so there is now only a 60% chance of success when trying to cancel. Dipping a phoenix egg in a potion of amnesia will always sterilize the egg however.
Polymorph needed to be addressed as well, because one could just zap a phoenix with polymorph and be rid of the pest. So now, when a phoenix doesn't naturally resist a polymorph, it will violently resist and explode, always reverting to it's egg form. This also entailed protecting phoenix eggs from polymorph, because without that protection, the polymorph beam would always poly the egg immediately after the phoenix exploded.

Lastly, in order to avoid some weird philosophical issues and gameplay paradoxes, I added the NOPOLY tag to the phoenix so the player can never take their form.

Thanks to Noisytoot for inspiring these ideas with the observation that phoenix eggs could simply be placed in a container (similar to zombie corpses). Phoenixes should have a special place in Hack'EM so they are revered and despised, plus they add significant value to potions of amnesia.