Rope bridge levels from SpliceHack. From SpliceHack commit 26eee595d. I've been hoping to add these for a while, I think they'll add some really nice atmosphere to the latter half of the main dungeon.
In SpliceHack, the chasms and bridges would start appearing after level 3, but we're going to delay them until at least level 15. This coincides with the quests and signals a change in the difficulty of the dungeon. Also, starting at level 25, the chasms can start being composed of pools or lava, adding more danger as the player advances.

This commit also includes the follow-up commit 04b97eed45, which fixes some initial bugs with the bridges and generation.