New monster: arch-vile (from DOOM). True to it's origins, this monster is a high-level mage which has two core mechanics derived from DOOM.
The first is specializing in the fire pillar attack. This is a clerical spell, however, the arch-vile will only cast this spell rather than having access to the entire clerical library. One exception is when the arch-vile is low on health, it will also cast self healing. The fire pillar has been adjusted so it can be used at range, before it was only usable within melee range.

The second core mechanic is the ability to ressurect corpses. I was actually able to reuse the existing infestation mechanic that moldiers introduced. When arch-viles walk over any corpse, they can ressurect it.

With their blazing speed and skittish behavior, I predict these will be quite a force for the player to deal with and likely become a popular genocide target. I debated whether to give them a melee/touch attack or any other attacks, but I think the two main mechanics combined with their speed and evasive movement will be quite enough.