More Shamblestick finessing.
Added support for many more random special attack types, including drain life and beheading. I improved how the to-hit and damage bonuses were applied. If there are few abilities, there is are high to-hit/damage bonuses. If there are many abilities, then the bonuses are small. If the player is lucky enough to get a 0 for the damage bonus, it naturally becomes a double damage weapon. If Shamblestick gets a bane, it will never be against the player's race. I removed the cross-aligned bane property because it could make it unusable if it didn't match the player's alignment. I also updated the adtyp_str function to be easier to use and understand.