Imported deck of fate and playing card deck from SpliceHack.
I made lots of tweaks to the deck of fate. I want this to be a pseudo-tribute to the gypsy, so I tried to use the best effects of both the SLASH'EM gypsy tarot and SpliceHack's deck of fate. Where possible, I tried to use the less severe effect or an effect that made more sense thematically.

A notable exception is the Death card, which will inflict insta-death. However, I was able to fit in the recent 3.7 saving grace mechanic. If the player hasn't used their saving grace yet and if they have positive Luck then they will spared once.

For now the tower effect will reveal the location of Vlad's Tower, but I would like to extend it a bit more to show the Wizard's tower.

In SpliceHack there is an interesting interaction with demons stealing your items and playing games with them. For now, I'm going to put this on the backburner with the intention of investigating it later and bringing it back with refinements.

The deck of cards is fairly simple but useful - it will reveal your current luck standing.