For cartomancers, extend the spell being mechanics to other sources of creating monsters. If a monster is created via scroll/wand/spell of create monster (or almost any magical source), it will instead be temporary spell being with a pre-determined lifetime. I think this makes for some interesting interactions with the cartomancer. It's also the way monster creation works in FIQhack but here we get to limit it to only one role.
The easiest way to implement this was to modify create_critters. This also effects many other ways monsters are generated, so it might be worth revisiting to limit the spell beings to only magical sources, but for now it seems okay.

While handling the muse cases, I also considered converting the monster spells of make undead and summon nasties, but this will require a little more work. I need to do a follow-up commit that makes sure the birth count of monsters that are spell-beings are not incremented.